moved RingBuffer to SerialMock
