--- test/mocks/RingBuffer.cpp
@@ -1,1 +1,0 @@
-#include "RingBuffer.h"
